[refactor][볼링-step3] frame 패키지 정리
